extract tag_keys and convert
