+++ convert/convert.go
@@ -1,9 +1,14 @@
 package convert
 
 import (
+	"encoding/json"
+	"github.com/env0/terratag/errors"
+	"github.com/env0/terratag/tag_keys"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/thoas/go-funk"
+	"github.com/zclconf/go-cty/cty"
+	"log"
 	"strings"
 )
 
@@ -78,4 +83,87 @@ func stringifyExpression(tokens hclwrite.Tokens) string {
 	expression = strings.TrimSuffix(expression, "${")
 
 	return expression
+}
+
+func AppendLocalsBlock(file *hclwrite.File, filename string, terratag TerratagLocal) {
+	file.Body().AppendNewline()
+	locals := file.Body().AppendNewBlock("locals", nil)
+	file.Body().AppendNewline()
+
+	locals.Body().SetAttributeValue(tag_keys.GetTerratagAddedKey(filename), cty.StringVal(terratag.Added))
+}
+
+func AppendTagBlocks(resource *hclwrite.Block, tags string) {
+	var tagsMap map[string]string
+	err := json.Unmarshal([]byte(tags), &tagsMap)
+	errors.PanicOnError(err, nil)
+
+	for key, value := range tagsMap {
+		resource.Body().AppendNewline()
+		tagBlock := resource.Body().AppendNewBlock("tag", nil)
+		tagBlock.Body().SetAttributeValue("key", cty.StringVal(key))
+		tagBlock.Body().SetAttributeValue("value", cty.StringVal(value))
+		tagBlock.Body().SetAttributeValue("propagate_at_launch", cty.BoolVal(true))
+	}
+}
+
+func UnquoteTagsAttribute(swappedTagsStrings []string, text string) string {
+	for _, swappedTagString := range swappedTagsStrings {
+		escapedByWriter := strings.ReplaceAll(swappedTagString, "\"", "\\\"")
+
+		if strings.HasPrefix(swappedTagString, "${") && strings.HasSuffix(swappedTagString, "}") {
+			escapedByWriter = strings.ReplaceAll(escapedByWriter, "${", "$${")
+		} else {
+			escapedByWriter = "\"" + escapedByWriter + "\""
+		}
+
+		text = strings.ReplaceAll(text, escapedByWriter, swappedTagString)
+	}
+	return text
+}
+
+func MoveExistingTags(filename string, terratag TerratagLocal, resource *hclwrite.Block) bool {
+	var existingTags hclwrite.Tokens
+
+	// First we try to find tags as attribute
+	tagsAttribute := resource.Body().GetAttribute("tags")
+
+	if tagsAttribute != nil {
+		// If attribute found, get its value
+		log.Print("Preexisting tags ATTRIBUTE found on resource. Merging.")
+		existingTags = tagsAttribute.Expr().BuildTokens(hclwrite.Tokens{})
+	} else {
+		// Otherwise, we try to get tags as block
+		tagsBlock := resource.Body().FirstMatchingBlock("tags", nil)
+		if tagsBlock != nil {
+			quaotedTagBlock := quoteBlockKeys(tagsBlock)
+			existingTags = funk.Tail(quaotedTagBlock.BuildTokens(hclwrite.Tokens{})).(hclwrite.Tokens)
+			// If we did get tags from block, we will now remove that block, as we're going to add a merged tags ATTRIBUTE
+			removeBlockResult := resource.Body().RemoveBlock(tagsBlock)
+			if removeBlockResult == false {
+				log.Fatal("Failed to remove found tags block!")
+			}
+		}
+	}
+
+	if existingTags != nil {
+		terratag.Found[tag_keys.GetResourceExistingTagsKey(filename, resource)] = existingTags
+		return true
+	}
+	return false
+}
+
+func quoteBlockKeys(tagsBlock *hclwrite.Block) *hclwrite.Block {
+	// In HCL, block keys must NOT be quoted
+	// But we need them to be, as we throw them into a map() function as strings
+	quaotedTagBlock := hclwrite.NewBlock(tagsBlock.Type(), tagsBlock.Labels())
+	for key, value := range tagsBlock.Body().Attributes() {
+		quaotedTagBlock.Body().SetAttributeRaw("\""+key+"\"", value.Expr().BuildTokens(hclwrite.Tokens{}))
+	}
+	return quaotedTagBlock
+}
+
+type TerratagLocal struct {
+	Found map[string]hclwrite.Tokens
+	Added string
 }
