+++ main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/env0/terratag/convert"
 	. "github.com/env0/terratag/errors"
 	"github.com/env0/terratag/file"
+	"github.com/env0/terratag/tag_keys"
 	. "github.com/env0/terratag/terraform"
 	. "github.com/env0/terratag/tfschema"
 	"github.com/hashicorp/hcl/v2/hclwrite"
@@ -49,7 +50,7 @@ func tagFileResources(path string, dir string, tags string, tfVersion int) {
 	log.Print("Processing file ", path)
 	hcl := file.ReadHCLFile(path)
 
-	terratag := TerratagLocal{
+	terratag := convert.TerratagLocal{
 		Found: map[string]hclwrite.Tokens{},
 		Added: jsonToHclMap(tags),
 	}
@@ -73,7 +74,7 @@ func tagFileResources(path string, dir string, tags string, tfVersion int) {
 					// https://www.terraform.io/docs/providers/aws/r/autoscaling_group.html
 					swappedTagsStrings = append(swappedTagsStrings, tagResource(filename, terratag, resource, tfVersion))
 				} else {
-					appendTagBlocks(resource, tags)
+					convert.AppendTagBlocks(resource, tags)
 				}
 				anyTagged = true
 			} else {
@@ -83,12 +84,12 @@ func tagFileResources(path string, dir string, tags string, tfVersion int) {
 	}
 
 	if anyTagged {
-		appendLocalsBlock(hcl, filename, terratag)
+		convert.AppendLocalsBlock(hcl, filename, terratag)
 
 		text := string(hcl.Bytes())
 
 		swappedTagsStrings = append(swappedTagsStrings, terratag.Added)
-		text = unquoteTagsAttribute(swappedTagsStrings, text)
+		text = convert.UnquoteTagsAttribute(swappedTagsStrings, text)
 
 		file.ReplaceWithTerratagFile(path, text)
 	} else {
@@ -96,16 +97,16 @@ func tagFileResources(path string, dir string, tags string, tfVersion int) {
 	}
 }
 
-func tagResource(filename string, terratag TerratagLocal, resource *hclwrite.Block, tfVersion int) string {
+func tagResource(filename string, terratag convert.TerratagLocal, resource *hclwrite.Block, tfVersion int) string {
 	log.Print("Resource taggable, processing...")
 
-	hasExistingTags := moveExistingTags(filename, terratag, resource)
+	hasExistingTags := convert.MoveExistingTags(filename, terratag, resource)
 
 	tagsValue := ""
 	if hasExistingTags {
-		tagsValue = "merge( " + convert.GetExistingTagsExpression(terratag.Found[getResourceExistingTagsKey(filename, resource)]) + ", local." + getTerratagAddedKey(filename) + ")"
+		tagsValue = "merge( " + convert.GetExistingTagsExpression(terratag.Found[tag_keys.GetResourceExistingTagsKey(filename, resource)]) + ", local." + tag_keys.GetTerratagAddedKey(filename) + ")"
 	} else {
-		tagsValue = "local." + getTerratagAddedKey(filename)
+		tagsValue = "local." + tag_keys.GetTerratagAddedKey(filename)
 	}
 
 	if tfVersion == 11 {
@@ -115,28 +116,6 @@ func tagResource(filename string, terratag TerratagLocal, resource *hclwrite.Blo
 	resource.Body().SetAttributeValue("tags", cty.StringVal(tagsValue))
 
 	return tagsValue
-}
-
-func appendLocalsBlock(file *hclwrite.File, filename string, terratag TerratagLocal) {
-	file.Body().AppendNewline()
-	locals := file.Body().AppendNewBlock("locals", nil)
-	file.Body().AppendNewline()
-
-	locals.Body().SetAttributeValue(getTerratagAddedKey(filename), cty.StringVal(terratag.Added))
-}
-
-func appendTagBlocks(resource *hclwrite.Block, tags string) {
-	var tagsMap map[string]string
-	err := json.Unmarshal([]byte(tags), &tagsMap)
-	PanicOnError(err, nil)
-
-	for key, value := range tagsMap {
-		resource.Body().AppendNewline()
-		tagBlock := resource.Body().AppendNewBlock("tag", nil)
-		tagBlock.Body().SetAttributeValue("key", cty.StringVal(key))
-		tagBlock.Body().SetAttributeValue("value", cty.StringVal(value))
-		tagBlock.Body().SetAttributeValue("propagate_at_launch", cty.BoolVal(true))
-	}
 }
 
 func jsonToHclMap(tags string) string {
@@ -149,74 +128,4 @@ func jsonToHclMap(tags string) string {
 		mapContent = append(mapContent, "\""+key+"\"="+"\""+value+"\"")
 	}
 	return "{" + strings.Join(mapContent, ",") + "}"
-}
-
-func unquoteTagsAttribute(swappedTagsStrings []string, text string) string {
-	for _, swappedTagString := range swappedTagsStrings {
-		escapedByWriter := strings.ReplaceAll(swappedTagString, "\"", "\\\"")
-
-		if strings.HasPrefix(swappedTagString, "${") && strings.HasSuffix(swappedTagString, "}") {
-			escapedByWriter = strings.ReplaceAll(escapedByWriter, "${", "$${")
-		} else {
-			escapedByWriter = "\"" + escapedByWriter + "\""
-		}
-
-		text = strings.ReplaceAll(text, escapedByWriter, swappedTagString)
-	}
-	return text
-}
-
-func moveExistingTags(filename string, terratag TerratagLocal, resource *hclwrite.Block) bool {
-	var existingTags hclwrite.Tokens
-
-	// First we try to find tags as attribute
-	tagsAttribute := resource.Body().GetAttribute("tags")
-
-	if tagsAttribute != nil {
-		// If attribute found, get its value
-		log.Print("Preexisting tags ATTRIBUTE found on resource. Merging.")
-		existingTags = tagsAttribute.Expr().BuildTokens(hclwrite.Tokens{})
-	} else {
-		// Otherwise, we try to get tags as block
-		tagsBlock := resource.Body().FirstMatchingBlock("tags", nil)
-		if tagsBlock != nil {
-			quaotedTagBlock := quoteBlockKeys(tagsBlock)
-			existingTags = funk.Tail(quaotedTagBlock.BuildTokens(hclwrite.Tokens{})).(hclwrite.Tokens)
-			// If we did get tags from block, we will now remove that block, as we're going to add a merged tags ATTRIBUTE
-			removeBlockResult := resource.Body().RemoveBlock(tagsBlock)
-			if removeBlockResult == false {
-				log.Fatal("Failed to remove found tags block!")
-			}
-		}
-	}
-
-	if existingTags != nil {
-		terratag.Found[getResourceExistingTagsKey(filename, resource)] = existingTags
-		return true
-	}
-	return false
-}
-
-func quoteBlockKeys(tagsBlock *hclwrite.Block) *hclwrite.Block {
-	// In HCL, block keys must NOT be quoted
-	// But we need them to be, as we throw them into a map() function as strings
-	quaotedTagBlock := hclwrite.NewBlock(tagsBlock.Type(), tagsBlock.Labels())
-	for key, value := range tagsBlock.Body().Attributes() {
-		quaotedTagBlock.Body().SetAttributeRaw("\""+key+"\"", value.Expr().BuildTokens(hclwrite.Tokens{}))
-	}
-	return quaotedTagBlock
-}
-
-func getTerratagAddedKey(filname string) string {
-	return "terratag_added_" + filname
-}
-
-func getResourceExistingTagsKey(filename string, resource *hclwrite.Block) string {
-	delimiter := "__"
-	return "terratag_found_" + filename + delimiter + strings.Join(resource.Labels(), delimiter)
-}
-
-type TerratagLocal struct {
-	Found map[string]hclwrite.Tokens
-	Added string
 }
